Playing with dataStore.
Discrepencies between the types of stateflows used is causing problems.
The codelab didn't use mutable stateflows - and that has been
confusing. I am doing something wrong with their uiState template.
They aren't initializing the uiState based on a mutable state
and are simply using a static uiState that's public and exposed
to the compose funcitons.