Adding datastore functionality.
This commit SEEMS to be a stable point in saving data store value for
desired monthly spending.

The part that was a little more arduous about this was the working with
flows. How could I successfully work with one flow (in datastore) and take
advantage of the flow to link to the mutable state flow of the viewmodel.